requirements.txt
Django==3.2.18
Pillow
environs==9.5.0
django-admin-sortable==2.2.4
django-tinymce==3.6.1
